Final code

+++ src/main/java/com/appleindustries/challenge/repositories/model/PackageType.java
@@ -19,5 +19,5 @@ public class PackageType {
 
     String name;
 
-    int price;
+    Double price;
 }
